% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{blockr.fredr-package}
\alias{blockr.fredr}
\alias{blockr.fredr-package}
\title{blockr.fredr: FRED data blocks for 'blockr.core'}
\description{
Provides a data-import block backed by the 'fredr' package for 'blockr.core'.
}
\author{
\strong{Maintainer}: Christoph Sax \email{christoph@cynkra.com}

}
\keyword{internal}
