% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fred-block.R
\name{register_fredr_blocks}
\alias{register_fredr_blocks}
\title{Register FRED blocks}
\usage{
register_fredr_blocks()
}
\description{
Registers all FRED blocks with the blockr.core registry.
This function is called automatically when the package is loaded.
}
